shared-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
nemoguardrails>=0.7.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-core>=0.1.0
pyyaml>=6.0